[front] Reuse skill details in enable_skill actions
